+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,15 +2,15 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_199418: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_199217: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_199416: (a: number, b: number, c: number, d: number) => void;
-export const __wasm_bindgen_func_elem_199430: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_192594: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_192605: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_199420: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_194719: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_194739: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199454: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_199253: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199456: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_199452: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_199466: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_192630: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_192641: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_194755: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_194775: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;